requirements.txt
django
django-compat